+++ packages/ui/Marquee/MarqueeDecorator.js
@@ -374,6 +374,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (this.context && this.context.register) {
 				this.sync = true;
 				this.context.register(this, {
+					restart: this.restart,
 					start: this.start,
 					stop: this.stop
 				});
@@ -412,7 +413,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				);
 
 				this.invalidateMetrics();
-				this.cancelAnimation();
+				this.cancelAnimation(forceRestartMarquee);
 				if (forceRestartMarquee && marqueeOn === 'focus') {
 					this.resetAnimation();
 				}
@@ -704,6 +705,15 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		};
 
 		/*
+		 * Restarts the animation
+		 *
+		 * @returns	{undefined}
+		 */
+		restart = () => {
+			this.restartAnimation();
+		};
+
+		/*
 		 * Starts marquee animation with synchronization, if not already animating and a timer is
 		 * not already active to start.
 		 *
@@ -747,7 +757,7 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		 *
 		 * @returns {undefined}
 		 */
-		restartAnimation = (delay) => {
+		restartAnimation = (delay = MINIMUM_MARQUEE_RESET_DELAY) => {
 			flushSync(() => {
 				this.setState({
 					animating: false
@@ -795,9 +805,9 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		 *
 		 * @returns {undefined}
 		 */
-		cancelAnimation = () => {
+		cancelAnimation = (retryStartingAnimation = false) => {
 			if (this.sync) {
-				this.context.cancel(this);
+				this.context.cancel(retryStartingAnimation);
 				return;
 			}