+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,9 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Changed
 
 - `ui/Marquee.MarqueeDecorator` to use `@enact/i18n/utils.isRtlText` API to determine the text directionality
+### Fixed
+
+- `ui/Marquee.MarqueeDecorator` to start animation properly when synchronized by `ui/Marquee.MarqueeController` and text changed
 
 ## [4.9.0-beta.1] - 2024-06-17