+++ packages/ui/CHANGELOG.md
@@ -14,6 +14,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 - `ui/SlotItem`
 - `ui/ToggleItem`
 - `ui/VirtualList.VirtualListBase`
+
 ### Changed
 
 - `ui/Marquee.MarqueeDecorator` to use `@enact/i18n/utils.isRtlText` API to determine the text directionality